refactor(release): fix paths for the src/ layout
The mechanical consequences of the move, exactly as scoped on #335:
cross-folder imports go through ../lib and ../checks; workspaceRoot
gains two .. segments (src/lib is two levels deeper); the mise release
task run lines and the packed-manifest gate point at src/steps and
src/checks; package scripts and the node --test glob move to src/**;
the workspace.ts export retargets its file so vue-check's
release/workspace.ts specifier is untouched; two workflow comments
naming old paths updated. No behavior changes.
