+++ tools/release/package.json
@@ -5,16 +5,16 @@
   "description": "Owns the release-pipeline logic: the tested entries the publish and bump-version workflows invoke",
   "type": "module",
   "exports": {
-    "./workspace.ts": "./workspace.ts"
+    "./workspace.ts": "./src/lib/workspace.ts"
   },
   "scripts": {
-    "check:pack": "node check-pack.ts",
-    "check:published-diff": "node check-published-diff.ts",
+    "check:pack": "node src/checks/check-pack.ts",
+    "check:published-diff": "node src/checks/check-published-diff.ts",
     "format": "oxfmt \"**/*.{ts,js,json,jsonc,md}\"",
     "format:check": "oxfmt --check \"**/*.{ts,js,json,jsonc,md}\"",
     "lint": "oxlint .",
-    "resolve:published": "node resolve-published.ts",
-    "test": "node --test \"*.test.ts\"",
+    "resolve:published": "node src/checks/resolve-published.ts",
+    "test": "node --test \"src/**/*.test.ts\"",
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "devDependencies": {
